Improved screen for newly added language tableView also made in MVC.
