+++ VocabularyTrainer/Controller/TrainingViewController.swift
