+++ VocabularyTrainer/Model/Extensions.swift
