+++ VocabularyTrainer/Model/Helper.swift
