+++ VocabularyTrainer/Model/de.lproj/Localizable.strings
