+++ VocabularyTrainer/Model/en.lproj/Localizable.strings
