+++ VocabularyTrainer/View/LanguageScreenViewController/VocabularyCell.swift
